# backend/openapi/flow-api.yaml
openapi: 3.0.3
info:
  title: Flow Service API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/stages/{stageId}/fields:
    get:
      summary: List fields for a stage
      parameters:
        - in: path
          name: stageId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    put:
      summary: Replace fields for a stage
      parameters:
        - in: path
          name: stageId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  label: { type: string }
                  type: { type: string }
                  required: { type: boolean }
                  placeholder: { type: string }
                  options:
                    type: array
                    items: { type: string }
                  order: { type: integer }
      responses:
        '200': { description: OK }
...